Extract batched prediction logic to 'PredictionUtils.py' to avoid dependencies on matplotlib, torchvision, etc.
